python_app/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Admin==2.0.2
Flask-Migrate==4.1.0
Pillow==12.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1